Refactor input question view
